perf(graph): CSR build stops allocating one String per edge; snapshot halves its index arrays
Three defects in the same file, fixed together because they share the
snapshot's layout:

- SnapshotBuilder::build interned labels with entry(label_str.to_string()):
  entry() demands an owned key, so every EDGE paid a String allocation to
  probe a map whose distinct-key count is the label vocabulary (typically
  tens). A 10M-edge rebuild burned 10M allocations to learn ~20 facts.
  The probe is now a &str get(); the owned Arc<str> is built only on the
  first sighting of a label.

- label_table (Vec) and label_to_idx (map keys) each owned a full copy of
  every label string. Both sides now share one Arc<str> per distinct
  label — the same fix StringTable just received.

- offsets was Vec<usize> and node_to_index carried usize values, though
  every value is bounded by the snapshot's edge count (a u32-overflowing
  snapshot would already hold >100 GB of parallel u64 arrays — the bound
  is debug-asserted at build). Halving offsets doubles the node ranges
  per cache line on the BFS hot path; the snapshot is rebuilt from
  scratch and never persisted, so no disk format is involved.

Validated: full lib suite 5352/5352 (csr 46/46, graph 582/582), clippy
pedantic at CI-exact flags, fmt, no-default-features -Dwarnings.

+++ crates/velesdb-core/src/collection/graph/csr_tests.rs
@@ -293,7 +293,7 @@ mod property_tests {
             }
 
             // offsets[M] == N (last offset == total edges)
-            prop_assert_eq!(*offsets.last().unwrap(), total_edges);
+            prop_assert_eq!(*offsets.last().unwrap() as usize, total_edges);
 
             // Per-node degree matches EdgeStore
             for &nid in &source_nodes {
